新しいMacで動作させるのに必要な変更を実施

+++ setup.sh
@@ -17,16 +17,23 @@ sudo cp malware_monitor_wrapper /usr/local/bin/malware_monitor_wrapper
 sudo chmod 700 /usr/local/bin/malware_monitor_wrapper
 sudo chown root:wheel /usr/local/bin/malware_monitor_wrapper
 sudo cp freshclam.conf /usr/local/etc/clamav/freshclam.conf
-sudo cp clamd.conf /usr/local/etc/clamav/clamd.conf
 sudo mkdir /usr/local/etc/clamav/quarantine 2> /dev/null
 sudo mkdir /var/log/clamav/ 2> /dev/null
 sudo mkdir -p /usr/local/var/run/clamav 2> /dev/null
-sudo chown _clamav /var/log/clamav/
+sudo chown -R _clamav /var/log/clamav/
 sudo mkdir -p /usr/local/Cellar/clamav/0.104.1/share/clamav
 sudo chown _clamav /usr/local/Cellar/clamav/*/share/clamav/
 sudo chown _clamav /usr/local/var/run/clamav
-sudo /usr/local/bin/freshclam -v
-sudo install -m 644 ./jp.aeyesec.Clamd.plist /Library/LaunchDaemons
+if [ -e "/opt/homebrew/sbin/clamd" ]; then
+  sudo chown _clamav /opt/homebrew/var/lib/clamav
+  sudo /opt/homebrew/bin/freshclam -v
+  sudo install -m 644 ./jp.aeyesec.Clamd2.plist /Library/LaunchDaemons/jp.aeyesec.Clamd.plist
+  sudo cp clamd.conf /opt/homebrew/etc/clamav/clamd.conf
+else
+  sudo /usr/local/bin/freshclam -v
+  sudo install -m 644 ./jp.aeyesec.Clamd.plist /Library/LaunchDaemons
+  sudo cp clamd.conf /usr/local/etc/clamav/clamd.conf
+fi
 sudo launchctl unload /Library/LaunchDaemons/jp.aeyesec.Clamd.plist 2> /dev/null
 sudo launchctl load /Library/LaunchDaemons/jp.aeyesec.Clamd.plist
 sudo install -m 644 ./jp.aeyesec.ClamdOnAccessAlert.plist /Library/LaunchDaemons
